main_status can be used in the GUI or CL
When instantiated in the GUI this module will write
to an text box, there are many modules that can use the
same reference (on the same thread) so using this will
allow users to send debug messages in a uniform way to
the same place

When not using the GUI (either using the CL version
of the tool or testing this module) can be instantiated
with the 'Dummy' or non GUI output. This output simply
goes to the command line.

Some features:

Specify the sevarity of the message and these messages can be filtered
to reduce the number of messages received. For example during Debug
designers can use the output.Debug(self, message) to write a lot of
verbose messages that are helpful when debugging. In order to not
show these messages the status output debug level can be set to
a higher value (like 'Info') and then Debug messages will not be
shown.