Changed cbuilder directory
cbuilder directory uses a new class called 'CBuilder in the cbuilder.py module.
The project class is loaded and unluaded outside the plugin's control so
functions that I attached to toolbar items didn't exist.

I modified wizard to allow users to select whether a core will be a memory
or a peripheral device

In the cbuilder.py I added functions that perform compilation, simulation and
waveform viewer.

I still need to fix the demo project