Implemented different folder organization and sim
New folder organization is simpiler to read because the rtl and the sim
are right next to each other as apposed to being in two separate trees

Implemented scons build!

scons controls the build process now. I needs to be done better (creating a
tool, using a global configuration file for the build process) but for now
it will build a core both within the IDE and outside it.

Fixed various bugs within wishbone_main and the interconnect. I'll need to
modify the interconnect generator in the future.