moved a lot of stuff and changed the scripts
Changed the structure of the directory from

nysa/cbuilder/rtl to nysa/cbuilder/verilog

it makes more sense because rtl is only synthesizable, which is what the
previous version had but in order to implement simulation I had to use
simlinks to connect the two unfortunately simlinks don't work across all
platforms. So this new directory structure handles this better by localizing
the simulation directory right next to the rtl.

the build structure primarily uses scons now instead of make. It is a lot easier
to use compared to Makefile... well there is a little growing pains with it
but it looks very promising