refactored the ibuilder project controller
Moved the controller for both wishbone and axi to the project directory
instead of within the editor. I realized the editor should be thought of
as a view and the controller should be separate.

I'm about to add a constraint editor which is a separate view of the
same project so it needs the same controller