implemented bus level signal/constraint comparison
Now when adding the signals to the signal view, the controller will
check if signals from a bus are bound.

If all signals on a bus are bound then the bus is not displayed in the
available signals.

If not all of the signals on a bus are bound then the bus is displayed
in the available signals, but in the future the signals will not be
enabled (grayed out)