hijacked the build file and build project
Unfortunately I can't use the build project and build file as they are
within Ninja because when I want to build a project the first thing
Ninja does is it looks for the current editor. I should have been able
to subclass editor but the problem was that the editor class is for
plain text editor, not a generic widget type editor.

That wouldn't be so bad but because I can't declare my widgets as editor
I don't receive any of the build signals, so what I did was disconnect
the run project, and run file signals and connected my own functions.
After I check to see that I am not trying to build my own functions I
pass the context onto the normal Ninja-ide builder