A lot of reworking this file
This file is way too big, it needs to be split up. since the new naming
convention where the port direction prepends port names, wires start
with w_'s and registers start with r_'s I've had a hard time finding all
the small bugs with the design, and this one was very hard. I added a
new fuction that will generate the io buffer

I changed the spacing to 4 instead of 2 for indents to align with PEP8