Implemented build console output for ibuilder
Now I can watcth the rediculous amount of messages comming from the
Xilnx build tools. To accomplish this I needed to implement a 'consol'
so-to-speak which taked over the native console output originally
designed for Python. Fortunately it wasn't that challenging because I
already solved a lot of these issues with the cbuilder console output