DESIGN CHANGE/BUG FIX: Spcification Change/Data Loss
Memory is now accesssed through a 32-bit address mode instead of a byte
address interface. This simplifies transactions with the host and
modules that access the memory directly.

The bug fixes is specific to the case where the user writes in a
discontinuous fashion to the memory. So instead of writing to ALL of the
memory data is written a small amount at a time. The write stack was
intermittently dropping data. This was associated with counting. I've
observed that I have more consistent results when I implement a
'counting up' approach as apposed to count down to zero. Not only do I
need to implement less checks (to see if the count is at 1 or 0) it will
requrie less resources in the FPGA to implement counting as apposed to
subtraction