BUG FIX Wishbone master ping mistake
Wow! Wishbone master didn't de-assert master_ready when a ping occurs.
Granted the ping happens so fast that it doesn't really matter if it
goes down in most situations but this should have gon down every time
there is a command sent to the master.

This might explain why Dionysus doesn't respond sometimes